--- eunbko/src/components/Main/Modal2.js
@@ -1,126 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import styled, { keyframes, css } from 'styled-components';
-import oc from 'open-color';
-import { HeaderButton } from 'lib';
-
-const fadeIn = keyframes`
-  from {
-    opacity: 0
-  }
-  to {
-    opacity: 1
-  }
-`;
-
-const fadeOut = keyframes`
-  from {
-    opacity: 1
-  }
-  to {
-    opacity: 0
-  }
-`;
-
-const slideUp = keyframes`
-  from {
-    transform: translateY(200px);
-  }
-  to {
-    transform: translateY(0px);
-  }
-`;
-
-const slideDown = keyframes`
-  from {
-    transform: translateY(0px);
-  }
-  to {
-    transform: translateY(200px);
-  }
-`;
-
-// 모달 딤
-const ModalWrapper = styled.div`
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgba(0, 0, 0, 0.8);
-`;
-
-const ModalContents = styled.div`
-  width: 320px;
-  padding: 1.5rem;
-  background: white;
-  border-radius: 2px;
-  text-align:center;
-  h3 {
-    margin: 0;
-    font-size: 1.7rem;
-  }
-  p {
-    font-size: 1.125rem;
-  }
-  
-`;
-
-
-const ButtonGroup = styled.div`
-  margin-top: 3rem;
-  display: flex;
-  justify-content: center;
-`;
-
-const ShortMarginButton = styled(HeaderButton)`
-  & + & {
-    margin-left: 0.5rem;
-  }
-`;
-
-export function Modal2({
-  title,
-  children,
-  confirmText,
-  cancelText,
-  onConfirm,
-  onCancel,
-  visible,
-}) {
-  const [animate, setAnimate] = useState(false);
-  const [localVisible, setLocalVisible] = useState(visible);
-  useEffect(() => {
-    // visible 값이 true -> false 가 되는 것을 감지
-    if (localVisible && !visible) {
-      setAnimate(true);
-      setTimeout(() => setAnimate(false), 250);
-    }
-    setLocalVisible(visible);
-  }, [localVisible, visible]);
-
-  if (!animate && !localVisible) return null;
-  return (
-    <ModalWrapper>
-      <ModalContents>
-        <h3>{title}</h3>
-        <p>{children}</p>
-        <ButtonGroup>
-          <ShortMarginButton color="gray" onClick={onCancel}>
-            {cancelText}
-          </ShortMarginButton>
-          <ShortMarginButton color="pink" onClick={onConfirm}>
-            {confirmText}
-          </ShortMarginButton>
-        </ButtonGroup>
-      </ModalContents>
-    </ModalWrapper>
-  );
-}
-
-Modal2.defaultProps = {
-  confirmText: '확인',
-  cancelText: '취소',
-};
